修复发布bug 初步添加评论功能

+++ src/main/java/lin/community/communtiy/controller/PublishController.java
@@ -21,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    String edit(@PathVariable (name = "id") Integer id,
+    String edit(@PathVariable (name = "id") Long id,
                 Model model)
     {
         QuestionDTO question = questionService.getById(id);
@@ -42,7 +42,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest servletRequest,
             Model model
     ) {
